requirements.txt
# Requisitos para FOCA-LITE (by m10sec)
python-docx==1.1.0
openpyxl==3.1.2
PyPDF2==3.0.1
prompt_toolkit==3.0.43
pandas==2.2.2
googlesearch-python==1.1.0
pyexiftool==0.5.6